A mobile ad sample with place holders I created
# Mobile-Ad-Map
I was playing around and teaching myself mobile ads so I created an
HTML file with a map.

Monday, February 23rd, 2015

Copyright© 2015 Gigliotti, Angela

This project is solely intended as an example and is not to be copied
and reused as it is for the sole use of the author.